begin to restructure `FilterArray` module, starting with `primitive`
