+++ vortex-array/src/arrays/filter/execute/mod.rs
@@ -1,16 +1,14 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
-//! Execution logic for FilterArray - filters canonical arrays by a mask.
+//! Execution logic for [`FilterArray`].
+//!
+//! The main entrypoint is [`execute_filter`] which filters any [`Canonical`] array.
 
-use vortex_buffer::Buffer;
-use vortex_buffer::BufferMut;
 use vortex_error::VortexExpect;
 use vortex_mask::Mask;
-use vortex_mask::MaskIter;
 
 use crate::Canonical;
-use crate::IntoArray;
 use crate::arrays::BoolArray;
 use crate::arrays::BoolVTable;
 use crate::arrays::DecimalArray;
@@ -22,22 +20,21 @@ use crate::arrays::ListViewArray;
 use crate::arrays::ListViewRebuildMode;
 use crate::arrays::ListViewVTable;
 use crate::arrays::NullArray;
-use crate::arrays::PrimitiveArray;
 use crate::arrays::StructArray;
 use crate::arrays::StructVTable;
 use crate::arrays::VarBinViewArray;
 use crate::arrays::VarBinViewVTable;
-use crate::arrays::filter::FilterArray;
 use crate::compute::FilterKernel;
-use crate::validity::Validity;
 
-/// TODO: replace usage of compute fn.
+mod primitive;
+
+// TODO(connor): Stop using the old compute kernels and move all code into this module.
 /// Filter a canonical array by a mask, returning a new canonical array.
-pub fn filter_canonical(canonical: Canonical, mask: &Mask) -> Canonical {
+pub(super) fn execute_filter(canonical: Canonical, mask: &Mask) -> Canonical {
     match canonical {
         Canonical::Null(a) => Canonical::Null(filter_null(&a, mask)),
         Canonical::Bool(a) => Canonical::Bool(filter_bool(&a, mask)),
-        Canonical::Primitive(a) => Canonical::Primitive(filter_primitive(&a, mask)),
+        Canonical::Primitive(a) => Canonical::Primitive(primitive::filter_primitive(&a, mask)),
         Canonical::Decimal(a) => Canonical::Decimal(filter_decimal(&a, mask)),
         Canonical::VarBinView(a) => Canonical::VarBinView(filter_varbinview(&a, mask)),
         Canonical::List(a) => Canonical::List(filter_listview(&a, mask)),
@@ -57,60 +54,6 @@ fn filter_bool(array: &BoolArray, mask: &Mask) -> BoolArray {
         .vortex_expect("filter bool array")
         .as_::<BoolVTable>()
         .clone()
-}
-
-fn filter_primitive(array: &PrimitiveArray, mask: &Mask) -> PrimitiveArray {
-    use vortex_dtype::match_each_native_ptype;
-
-    // Lazy validity: wrap in FilterArray instead of eagerly filtering
-    let validity = match array.validity().vortex_expect("primitive validity") {
-        v @ (Validity::NonNullable | Validity::AllValid | Validity::AllInvalid) => v,
-        Validity::Array(arr) => {
-            Validity::Array(FilterArray::new(arr.clone(), mask.clone()).into_array())
-        }
-    };
-
-    match_each_native_ptype!(array.ptype(), |T| {
-        let filtered = filter_slice(
-            array.as_slice::<T>(),
-            mask,
-            FILTER_SLICES_SELECTIVITY_THRESHOLD,
-        );
-        PrimitiveArray::new(filtered, validity)
-    })
-}
-
-/// Threshold for choosing between indices vs slices filtering strategy.
-pub const FILTER_SLICES_SELECTIVITY_THRESHOLD: f64 = 0.8;
-
-/// Filter a typed buffer by a mask, returning a new buffer with only the selected elements.
-///
-/// This is the core filtering operation used by both FilterArray execution and the
-/// FilterKernel for primitive arrays.
-///
-/// # Arguments
-/// * `values` - The source slice of values to filter
-/// * `mask` - The mask indicating which elements to keep (must have `values()` available)
-/// * `selectivity_threshold` - Threshold for choosing between indices vs slices strategy
-pub fn filter_slice<T: Copy>(values: &[T], mask: &Mask, selectivity_threshold: f64) -> Buffer<T> {
-    let mask_values = mask
-        .values()
-        .vortex_expect("AllTrue and AllFalse should be handled by caller");
-
-    match mask_values.threshold_iter(selectivity_threshold) {
-        MaskIter::Indices(indices) => indices
-            .iter()
-            .copied()
-            .map(|idx| *unsafe { values.get_unchecked(idx) })
-            .collect(),
-        MaskIter::Slices(slices) => {
-            let mut output = BufferMut::with_capacity(mask.true_count());
-            for (start, end) in slices.iter().copied() {
-                output.extend_from_slice(&values[start..end]);
-            }
-            output.freeze()
-        }
-    }
 }
 
 fn filter_decimal(array: &DecimalArray, mask: &Mask) -> DecimalArray {
