+++ vortex-array/src/arrays/filter/mod.rs
@@ -2,12 +2,11 @@
 // SPDX-FileCopyrightText: Copyright the Vortex contributors
 
 mod array;
-mod execute;
-mod rules;
-mod vtable;
+pub use array::FilterArray;
 
-pub use array::*;
-pub use execute::FILTER_SLICES_SELECTIVITY_THRESHOLD;
-pub use execute::filter_canonical;
-pub use execute::filter_slice;
-pub use vtable::*;
+mod execute;
+
+mod rules;
+
+mod vtable;
+pub use vtable::FilterVTable;
