+++ vortex-array/src/arrays/primitive/compute/mod.rs
@@ -4,7 +4,6 @@
 mod between;
 mod cast;
 mod fill_null;
-mod filter;
 mod is_constant;
 mod is_sorted;
 mod mask;
